Add props to CardVehicleInfo and refactor the component styles
