+++ styles/CarVehicleInfo.js
@@ -1,0 +1,25 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  cardContainer: {
+    borderColor: "#3E8E7E",
+    borderWidth: 1,
+    width: 327,
+    height: 71,
+    borderRadius: 18,
+  },
+  modelText: {
+    fontWeight: "700",
+    fontSize: 19,
+    marginBottom: 8,
+    marginTop: 8,
+    marginLeft: 24,
+  },
+  carInfoText: {
+    alignSelf: "center",
+    fontWeight: "400",
+    fontSize: 16,
+  },
+});
+
+export default styles;
